+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/DependencyManager.groovy
@@ -2,6 +2,8 @@ package life.qbic.portal.offermanager
 
 import groovy.util.logging.Log4j2
 import life.qbic.business.offers.fetch.FetchOffer
+import life.qbic.business.products.archive.ArchiveProduct
+import life.qbic.business.products.create.CreateProduct
 import life.qbic.business.projects.create.CreateProject
 import life.qbic.datamodel.dtos.business.AcademicTitle
 import life.qbic.datamodel.dtos.business.AffiliationCategory
@@ -20,10 +22,12 @@ import life.qbic.portal.offermanager.components.offer.overview.projectcreation.C
 import life.qbic.portal.offermanager.components.person.search.SearchPersonView
 import life.qbic.portal.offermanager.components.person.search.SearchPersonViewModel
 import life.qbic.portal.offermanager.components.person.update.UpdatePersonViewModel
+import life.qbic.portal.offermanager.components.product.MaintainProductsPresenter
 import life.qbic.portal.offermanager.components.product.MaintainProductsView
 import life.qbic.portal.offermanager.components.product.MaintainProductsViewModel
 import life.qbic.portal.offermanager.components.product.create.CreateProductView
 import life.qbic.portal.offermanager.components.product.create.CreateProductViewModel
+import life.qbic.portal.offermanager.components.product.MaintainProductsController
 import life.qbic.portal.offermanager.dataresources.persons.AffiliationResourcesService
 import life.qbic.portal.offermanager.dataresources.persons.PersonDbConnector
 import life.qbic.portal.offermanager.dataresources.persons.CustomerResourceService
@@ -93,6 +97,7 @@ class DependencyManager {
     private SearchPersonViewModel searchPersonViewModel
     private CreatePersonViewModel createCustomerViewModelNewOffer
     private MaintainProductsViewModel maintainProductsViewModel
+    private MaintainProductsViewModel maintainProductsViewModelArchive
     private CreateProductViewModel createProductViewModel
     private CreateProductViewModel copyProductViewModel
     private CreateProjectViewModel createProjectModel
@@ -105,6 +110,8 @@ class DependencyManager {
     private CreateOfferPresenter createOfferPresenter
     private CreateOfferPresenter updateOfferPresenter
     private OfferOverviewPresenter offerOverviewPresenter
+    private MaintainProductsPresenter createProductPresenter
+    private MaintainProductsPresenter archiveProductPresenter
     private CreateProjectPresenter createProjectPresenter
 
     private PersonDbConnector customerDbConnector
@@ -124,6 +131,8 @@ class DependencyManager {
     private FetchOffer fetchOfferOfferOverview
     private FetchOffer fetchOfferCreateOffer
     private FetchOffer fetchOfferUpdateOffer
+    private CreateProduct createProduct
+    private ArchiveProduct archiveProduct
 
     private CreatePersonController createCustomerController
     private CreatePersonController updateCustomerController
@@ -132,6 +141,7 @@ class DependencyManager {
     private CreateOfferController createOfferController
     private CreateOfferController updateOfferController
     private OfferOverviewController offerOverviewController
+    private MaintainProductsController maintainProductController
     private CreateProjectController createProjectController
 
     private CreatePersonView createCustomerView
@@ -334,6 +344,12 @@ class DependencyManager {
         }
 
         try {
+            this.maintainProductsViewModelArchive = new MaintainProductsViewModel(productsResourcesService)
+        }catch (Exception e) {
+            log.error("Unexpected exception during ${MaintainProductsViewModel.getSimpleName()} view model setup.", e)
+        }
+
+        try {
             this.createProductViewModel = new CreateProductViewModel()
         }catch (Exception e) {
             log.error("Unexpected exception during ${CreateProductViewModel.getSimpleName()} view model setup.", e)
@@ -400,10 +416,21 @@ class DependencyManager {
         } catch (Exception e) {
             log.error("Unexpected exception during ${OfferOverviewPresenter.getSimpleName()} setup", e)
         }
+
+        try {
+            this.createProductPresenter = new MaintainProductsPresenter(this.maintainProductsViewModel, this.viewModel)
+        } catch (Exception e) {
+            log.error("Unexpected exception during ${MaintainProductsPresenter.getSimpleName()} setup", e)
+        }
+        try {
+            this.archiveProductPresenter = new MaintainProductsPresenter(this.maintainProductsViewModelArchive, this.viewModel)
+        } catch (Exception e) {
+            log.error("Unexpected exception during ${MaintainProductsPresenter.getSimpleName()} setup", e)
+        }
         try {
             this.createProjectPresenter = new CreateProjectPresenter(createProjectModel, viewModel)
         } catch (Exception e) {
-            log.error("Unexpected exception during ${OfferOverviewPresenter.getSimpleName()} setup", e)
+            log.error("Unexpected exception during ${CreateProjectPresenter.getSimpleName()} setup", e)
         }
     }
 
@@ -420,6 +447,9 @@ class DependencyManager {
         this.fetchOfferOfferOverview = new FetchOffer(offerDbConnector, offerOverviewPresenter)
         this.fetchOfferCreateOffer = new FetchOffer(offerDbConnector, createOfferPresenter)
         this.fetchOfferUpdateOffer = new FetchOffer(offerDbConnector, updateOfferPresenter)
+
+        this.createProduct = new CreateProduct(productsDbConnector,createProductPresenter)
+        this.archiveProduct = new ArchiveProduct(productsDbConnector,archiveProductPresenter)
 
         this.createProject = new CreateProject(createProjectPresenter, projectMainConnector, projectMainConnector)
     }
@@ -464,10 +494,17 @@ class DependencyManager {
         } catch (Exception e) {
             log.error("Unexpected exception during ${OfferOverviewController.getSimpleName()} setup", e)
         }
+
         try {
+            this.maintainProductController = new MaintainProductsController(this.createProduct, this.archiveProduct)
+        } catch (Exception e) {
+            log.error("Unexpected exception during ${MaintainProductsController.getSimpleName()} setup", e)
+        }
+
+        try{
             this.createProjectController = new CreateProjectController(this.createProject)
         } catch (Exception e) {
-            log.error("Unexpected exception during ${OfferOverviewController.getSimpleName()} setup", e)
+            log.error("Unexpected exception during ${CreateProjectController.getSimpleName()} setup", e)
         }
     }
 
@@ -555,23 +592,23 @@ class DependencyManager {
 
         CreateProductView createProductView
         try{
-            createProductView = new CreateProductView(createProductViewModel)
+            createProductView = new CreateProductView(createProductViewModel,maintainProductController)
         }catch(Exception e){
             log.error("Could not create ${CreateProductView.getSimpleName()} view.", e)
             throw e
         }
 
-        CreateProductView copyProductView
+        /**CreateProductView copyProductView
         try{
-            copyProductView = new CreateProductView(copyProductViewModel)
+            copyProductView = new CreateProductView(copyProductViewModel,maintainProductController)
         }catch(Exception e){
             log.error("Could not create ${CreateProductView.getSimpleName()} view.", e)
             throw e
-        }
+        }**/
 
         MaintainProductsView maintainProductsView
         try{
-            maintainProductsView = new MaintainProductsView(maintainProductsViewModel,createProductView,copyProductView)
+            maintainProductsView = new MaintainProductsView(maintainProductsViewModel,createProductView,maintainProductController)
         }catch (Exception e) {
             log.error("Could not create ${MaintainProductsView.getSimpleName()} view.", e)
             throw e