+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/product/MaintainProductsViewModel.groovy
@@ -1,5 +1,6 @@
 package life.qbic.portal.offermanager.components.product
 
+import groovy.beans.Bindable
 import life.qbic.datamodel.dtos.business.services.Product
 import life.qbic.portal.offermanager.dataresources.products.ProductsResourcesService
 
@@ -21,9 +22,9 @@ class MaintainProductsViewModel {
 
     ObservableList products = new ObservableList(new ArrayList<Product>())
 
-    Product selectedProduct
+    @Bindable Product selectedProduct
 
-    private final ProductsResourcesService productsResourcesService
+    final ProductsResourcesService productsResourcesService
 
     MaintainProductsViewModel(ProductsResourcesService productsResourcesService) {
         this.productsResourcesService = productsResourcesService
@@ -37,7 +38,12 @@ class MaintainProductsViewModel {
 
     private void subscribe(){
         productsResourcesService.subscribe({ product ->
-            products << product
+            refreshList()
         })
+    }
+
+    private void refreshList(){
+        products.clear()
+        products.addAll(productsResourcesService.iterator().toList())
     }
 }