+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
+
 1.0.0-alpha.5-SNAPSHOT (2021-03-17)
 --------------------------
 
@@ -12,6 +13,10 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 * Link offers to project now. The ``life.qbic.business.offers.Offer`` and ``life.qbic.portal
 .offermanager.dataresources.offers`` have been extended with a new property to associate it with
 an existing project by its project identifier.
+
+* Finalized the ``life.qbic.business.products.archive.ArchiveProduct`` and ``life/qbic/business/products/create/CreateProduct.groovy``
+use cases of the product maintenance and creation feature(`#411 <https://github.com/qbicsoftware/offer-manager-2-portlet/pull/411>`_).
+
 
 **Fixed**